+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Added an optional "Strict Mode" to SmartSelfie Enrollment and Authentication to achieve better pass rates. Set `useStrictMode=true` to enable this new, streamlined UI and associated active liveness tasks
 * Fixed a bug where some failed authentication requests were incorrectly handled
 * Fixed a bug where errors with no code were not being handled correctly
+* Fixed a bug on Selfie and Document capture success screen where the message was wrong
+* Fixed a bug where liveness files were missing on document verification jobs
 
 ## 10.1.6